MetaTransformer-Scripts/bioNN/logic_processing/logic-processing/core/solver/symbolic_solver.py
"""
Symbolic Logic Solver
Handles logical inference and problem-solving using Z3 SMT solver
"""

from pathlib import Path
from typing import Dict, List, Optional, Union
import z3
from loguru import logger
import yaml

class SymbolicSolver:
    """Symbolic logic solver using Z3"""
    
    def __init__(self, config_path: Optional[str] = None):
        self.config = self._load_config(config_path)
        self.solver = None
        self.context = None
        self._setup_solver()
    
    def _load_config(self, config_path: Optional[str]) -> Dict:
        """Load solver configuration"""
        if config_path is None:
            config_path = Path(__file__).parent.parent.parent / "config.yaml"
        
        with open(config_path, 'r') as f:
            full_config = yaml.safe_load(f)
            return full_config['logic_processing']['solver']
    
    def _setup_solver(self):
        """Initialize Z3 solver with configuration"""
        self.solver = z3.Solver()
        self.solver.set("timeout", self.config['timeout'] * 1000)  # Convert to milliseconds
        
        # Set optimization level
        z3.set_param('smt.random_seed', 42)  # For reproducibility
        z3.set_param('smt.relevancy', self.config['optimization_level'])
    
    def initialize(self) -> bool:
        """Initialize the solver"""
        try:
            # Test solver with a simple satisfiability check
            test_solver = z3.Solver()
            x = z3.Int('x')
            test_solver.add(x > 0)
            test_solver.add(x < 2)
            result = test_solver.check()
            return result == z3.sat
        except Exception as e:
            logger.error(f"Failed to initialize solver: {str(e)}")
            return False

    def create_predicates(self, predicate_defs: List[Dict]) -> Dict[str, z3.Function]:
        """Create Z3 function declarations for predicates"""
        predicates = {}
        for pred in predicate_defs:
            name = pred['name']
            arity = pred['arity']
            domain = [z3.DeclareSort('Object') for _ in range(arity)]
            predicates[name] = z3.Function(name, *domain, z3.BoolSort())
        return predicates

    def parse_logical_form(self, logical_form: str, predicates: Dict[str, z3.Function]) -> z3.BoolRef:
        """Parse logical formulation into Z3 expressions"""
        try:
            # Create parsing context
            ctx = {}
            for name, func in predicates.items():
                ctx[name] = func
                
            # Add logical operators to context
            ctx.update({
                'ForAll': z3.ForAll,
                'Exists': z3.Exists,
                'And': z3.And,
                'Or': z3.Or,
                'Not': z3.Not,
                'Implies': z3.Implies,
                'Iff': lambda x, y: z3.And(z3.Implies(x, y), z3.Implies(y, x))
            })
            
            # Parse and evaluate logical form
            # Note: This is a simplified example. In practice, you'd need a proper parser
            return eval(self._preprocess_logical_form(logical_form), ctx)
            
        except Exception as e:
            logger.error(f"Failed to parse logical form: {str(e)}")
            raise ValueError(f"Invalid logical form: {logical_form}")

    def _preprocess_logical_form(self, logical_form: str) -> str:
        """Preprocess logical form for Z3 compatibility"""
        # Replace logical symbols with Python/Z3 operations
        replacements = {
            '∀': 'ForAll',
            '∃': 'Exists',
            '∧': 'And',
            '∨': 'Or',
            '¬': 'Not',
            '→': 'Implies',
            '↔': 'Iff'
        }
        
        result = logical_form
        for symbol, replacement in replacements.items():
            result = result.replace(symbol, replacement)
            
        return result

    async def solve(self, logical_form: str, predicates: List[Dict]) -> Dict[str, Union[bool, str, Dict]]:
        """Solve a logical problem"""
        try:
            # Reset solver
            self.solver.reset()
            
            # Create predicates
            pred_dict = self.create_predicates(predicates)
            
            # Parse logical form
            formula = self.parse_logical_form(logical_form, pred_dict)
            
            # Add formula to solver
            self.solver.add(formula)
            
            # Check satisfiability
            result = self.solver.check()
            
            if result == z3.sat:
                model = self.solver.model()
                return {
                    "success": True,
                    "satisfiable": True,
                    "model": self._model_to_dict(model),
                    "confidence": 1.0
                }
            elif result == z3.unsat:
                return {
                    "success": True,
                    "satisfiable": False,
                    "reason": "Formula is unsatisfiable",
                    "confidence": 1.0
                }
            else:
                return {
                    "success": False,
                    "error": "Solver timeout or unknown result",
                    "confidence": 0.0
                }
                
        except Exception as e:
            logger.error(f"Solving failed: {str(e)}")
            return {
                "success": False,
                "error": str(e),
                "confidence": 0.0
            }

    def _model_to_dict(self, model: z3.ModelRef) -> Dict:
        """Convert Z3 model to dictionary"""
        result = {}
        for decl in model.decls():
            result[decl.name()] = str(model[decl])
        return result

    def verify_solution(self, solution: Dict[str, Union[bool, str, Dict]]) -> bool:
        """Verify the correctness of a solution"""
        try:
            if not solution['success']:
                return False
                
            # For satisfiable results, verify model consistency
            if solution.get('satisfiable', False):
                model = solution.get('model', {})
                # Add verification logic here
                return len(model) > 0
                
            return True
            
        except Exception as e:
            logger.error(f"Verification failed: {str(e)}")
            return False

if __name__ == "__main__":
    # Test the solver
    solver = SymbolicSolver()
    if solver.initialize():
        # Test with simple predicates
        test_predicates = [
            {"name": "Human", "arity": 1},
            {"name": "Mortal", "arity": 1}
        ]
        
        test_logic = "ForAll([x], Implies(Human(x), Mortal(x)))"
        
        import asyncio
        result = asyncio.run(solver.solve(test_logic, test_predicates))
        print(f"Solving result: {result}")
    else:
        print("Failed to initialize solver")