MetaTransformer-Scripts/bioNN/logic_processing/logic-processing/integration/quantum_bridge/connector.py
"""
Quantum Bridge Connector
Manages quantum coherence for logic processing integration
"""

import numpy as np
from dataclasses import dataclass
from typing import Dict, Optional
from loguru import logger
import yaml

@dataclass
class QuantumState:
    coherence: float
    entanglement: str
    pattern: str

class QuantumBridge:
    def __init__(self, config_path: Optional[str] = None):
        self.config = self._load_config(config_path)
        self.current_state = None
        self._initialize_bridge()
    
    def _load_config(self, config_path: Optional[str]) -> Dict:
        """Load quantum bridge configuration"""
        if config_path is None:
            config_path = "config.yaml"
            
        with open(config_path, 'r') as f:
            config = yaml.safe_load(f)
            return config['quantum_bridge']
    
    def _initialize_bridge(self):
        """Initialize quantum bridge parameters"""
        self.dimension_depth = self.config['dimension_depth']
        self.coherence_threshold = self.config['coherence_threshold']
        self.stability_index = self.config['stability_index']
        self.entanglement_pattern = self.config['entanglement_pattern']
        
        # Initialize quantum tensors
        self.coherence_matrix = np.eye(self.dimension_depth) * self.coherence_threshold
        self.stability_vector = np.ones(self.dimension_depth) * self.stability_index
    
    def connect(self) -> bool:
        """Establish quantum bridge connection"""
        try:
            self.current_state = QuantumState(
                coherence=self.coherence_threshold,
                entanglement=self.entanglement_pattern,
                pattern="initialized"
            )
            return True
        except Exception as e:
            logger.error(f"Failed to connect quantum bridge: {str(e)}")
            return False
    
    def synchronize_state(self, state_data: Dict) -> Dict:
        """Synchronize quantum state with new data"""
        try:
            # Update coherence matrix
            pattern_dimension = min(len(state_data), self.dimension_depth)
            update_mask = np.ones((pattern_dimension, pattern_dimension))
            self.coherence_matrix[:pattern_dimension, :pattern_dimension] *= update_mask
            
            return {
                "success": True,
                "coherence": float(np.mean(self.coherence_matrix)),
                "stability": float(np.mean(self.stability_vector))
            }
        except Exception as e:
            logger.error(f"State synchronization failed: {str(e)}")
            return {
                "success": False,
                "error": str(e)
            }
